refactor: 도메인 패키지 세분류. 자동차와 경주 분리 .
