+++ src/main/java/racing/domain/car/MovingStrategy.java
@@ -1,4 +1,4 @@
-package racing.domain;
+package racing.domain.car;
 
 public interface MovingStrategy {
     boolean isPossibleToMove();
