+++ src/main/java/racing/domain/car/RandomStrategy.java
@@ -1,4 +1,4 @@
-package racing.domain;
+package racing.domain.car;
 
 import java.util.Random;
