+++ src/main/java/racing/domain/race/RaceResult.java
@@ -1,4 +1,6 @@
-package racing.domain;
+package racing.domain.race;
+
+import racing.domain.car.Car;
 
 import java.util.ArrayList;
 import java.util.List;
