+++ src/test/java/racing/domain/RecordTest.java
@@ -2,6 +2,9 @@ package racing.domain;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racing.domain.car.AlwaysMovingStrategy;
+import racing.domain.car.Car;
+import racing.domain.race.Record;
 
 import static org.assertj.core.api.Assertions.assertThat;
